pyproject.toml
[project]
name = "dsa4264-geospatial-group6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "folium>=0.20.0",
    "geopandas>=1.1.3",
    "joblib>=1.5.3",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pydantic-ai>=1.76.0",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "shapely>=2.1.2",
    "statsmodels>=0.14.6",
    "uvicorn>=0.42.0",
]
